nextboot.sh: change default reboot behaviour to be more intuitive
The current implementation of `next_boot` would always reboot the system after selecting a new boot option, regardless of whether or not that was actually desired. This patch fixed that

This patch also cleans up some of the `echo` statements in the function and adds an error message if invalid arguments are passed.