Add fields to GPU log file - Update logging/gpulog.sh to include memory usage in the output file. - Modify nvidia-smi command in logging/gpulog.sh to include memory utilization and fan speed.
The git diff shows changes made to two files, logging/gpulog.sh and startup/init.sh. The main change is the addition of a new field 'memory_usage' in the output file for GPU logs. This was achieved by updating the command used to extract data from NVIDIA SMIs in logging/gpulog.sh. Additionally, the fan speed was also added as a new field. These changes were reflected in the commit message which describes the modifications made to both files.